requirements.txt
# Django Core
Django>=4.2,<5.0

# Django REST Framework
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0

# Swagger Documentation
drf-yasg>=1.21.7

# CORS
django-cors-headers>=4.3.0

# Database (SQLite for dev, uncomment for PostgreSQL)
# psycopg2-binary>=2.9.9

# Environment variables (optional)
# python-dotenv>=1.0.0

# Image handling
Pillow>=10.0.0

# Production (optional)
# gunicorn>=21.2.0
# whitenoise>=6.6.0
